First commit for the new KirHut Library Project.
libKirHut is intended to be a shared repository of code that KirHut applications use as a base from which the rest of the code base can be more easily fleshed out. Simply put, libKirHut is a set of common code that allows me to not write the same crap in the many multiple KirHut projects that are planned. I've wanted to do this for a while but could never settle on how, so this is a real attempt at making it work using Qt Creator and writing libKirHut to be primarily dynamically linked.

The goal is to make this library as flexible as possible, so I would like to support static building, building with and without Qt (so it should enhance the Qt stuff when it is there and ignore it otherwise), and heavily utilizing the C++ standard library. libKirHut is deliberately written to use modern C++20, just without modules yet due to the significant lack of good modules support across the build chain.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>