Mostly just adding everything for documentation: Doxyfile, tags, and all of that. The documentation was drafted and some more was added to how the global.hpp file worked.
NOTE: It is impossible to make documenting preprocessor defines not a pain in the ass. Preprocessor defines are documented with the source code document they are on, so you basically cannot use an external sheet with dummy information to perform the documentation. Worse, Doxygen does not generate documentation for preprocessor defines it does not find as active, so you have to do a bunch of hijinks just to put all of the preprocessor defines into scope for Doxygen, which in this case meant adding the KH_PRIV_DOCS define.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>