A Good Start to the Library for KirHut Application Development.
Well this has been an interesting experience. I've always wanted to have a library for myself or, rather, for KirHut to use as a base to develop applications from, so here's the beginnings of that library. There's honestly not much, but there are some basics that I have added for the purposes of testing this and being useful everywhere. The most important is actually the global and base headers, which have had their documentation heavily modified, but also exceptions.hpp, MD5 hashing, a ranges.hpp header, and the beginnings of the IO and logging systems have been put in place. The IO system seems to be in a reasonable state already, and uses Boost::nowide underneath it to provide UTF-8 IO on Windows, but the logging is basically non-existent and only an idea to use Loguru underneath to do the actual logging work, then provide an interface.

One of the big parts of this commit is the Documentation, which is now easy to generate with Doxygen. This commit includes the Doxygen-Awesome project, which creates a much more attractive set of Doxygen generated documentation (and will likely be how I always do this going forward). I have also added all of the tests that have so far been written, and I need to provide complete test coverage of what I have so far before I begin adding more functionality, namely before I get started on the logging functionality as that should be written using TDD. I also added a license and set it to the GNU GPL v3 for the entire library project. I was VERY tempted to use the AGPL for this as I am jealously defending this from competitors, but the GPL should be acceptable for the library itself as it does not provide much for competitive advantage and the GPL v3 is somewhat more liberal about what you can do with it, so maybe someone will make a web service using the library and never tell anyone.

A .clang-format file has been added, which is identical to all other KirHut software .clang-format files. The .clang-tidy and .clangd files will be updated when I know what to put in them. Started the README.md file, there is nothing special about it so far. This should be just about ready to upload to GitHub so that I can use it as a dependency in the other KirHut application projects!

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>