A whole bunch of uncommitted work needs to be committed, so here it is.
I... just can't fucking stand committing trash into a repo, but here I am because I cannot get this shit fucking together without testing it on the field. The entirety of the IO system, the new logging system which is just another library I ripped off, a bunch of work on the documentation (which isn't complete) and a totally haphazard bullshit system for argument parsing is being barely created as of this moment.

I fucking hate this committing shit. Please for the love of God just kill me.

This library will be the death of me. I need to get libKSC started too, which rolls in other dependencies (like requiring Qt instead of it just being optional) and FTXUI for TUI functionality. Then get both of the apps to rely on libKSC so I can actually fucking share code between the projects instead of copy/pasting every goddamn thing.

Signed-off-by: ZekeDragon <zeke@kirhut.com>