Added test for asBytes and asWritableBytes.
Good morning! This made me realize the two functions were broken. The functions are now not constexpr as asBytes/asWritableBytes makes no sense in a constexpr context whatsoever. The runtime representation of an object is unknowable at compile time, and constexpr contexts deliberately avoid doing anything with the actual runtime value of any set of data. Implementing this test also prompted me to create new toLittleEndian and toBigEndian functions, but those will need testing as well!

There is small polishing on documentation. Documentation work is ongoing and this will happen all the time.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>