Multiple tests for concepts in the base.hpp header added.
I had quite a time with this one. The concepts I was using, especially SpanOf, were fundamentally broken in some ways, so I had to work on how I was going to approach the design. I was forced to bifurcate Spanof into SpanOf and ReadableSpanOf to distinguish between spans you may want to write to vs spans you are only using for reading. They accomplish two different things, but are close. I added testing for the OneOf concept, ConvertsTo concept, and the InstanceOf concept, and updated all of their definitions to make sure the expected functionality is there.

Further commits will have tests for the rest of the concepts... ideally.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>