Testing for concepts in kh/base.hpp is completed.
Added multiple tests in the base testing application. Mostly all of them are static_asserts, so they all inherently pass if it builds, but this is important to make sure all of the library's concepts are properly working according to their intended function. The testing revealed that many of the concepts were inadequate or incorrectly implemented, and new implementations were drafted to ensure they would pass the testing.

This is why I need these tests so badly. I'm flying blind without anything to tell me if everything is working or not.

Also as a note, the tests are excessively boring, so I've been using an AI to generate them. It is one of the few things I've found it can do decently, though it still has several errors.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>