Finished Base file tests, cleaned up unused functions in base.hpp, cleaned documentation, and snuck in a fix for linker warnings with KH_EXPORT.
Finished all of the testing work for Base, and it turned out that of course the endian conversion functions were still broken. The testing proved this out and now the functions are working as expected. I also managed to find a way to get MSVC and GCC to stop issuing warnings when the KH_EXPORT macro was used to flag extern explicit template instances for linking. If it was in the header, MSVC warned and ignored, then failed because it needs to be in the source. If it is in the source, GCC warns and ignores, but everything builds because GCC gives no shits about the linker visibility. This was fixed with KH_EXPLICIT_TEMPLATE_EXPORT and KH_EXPLICIT_TEMPLATE_INSTANCE. Use the former on the extern explicit template declaration, and the latter on the definition line in the .cpp file.

Several unused functions were removed from base.hpp to alleviate maintenance burden and because I didn't want to bother with testing them. Documentation in a lot of areas was cleaned up a bit and expanded upon, mostly related to the endian functions.

As an aside, I also ran into a compiler bug on MSVC. I filed what happened here:
https://developercommunity.visualstudio.com/t/Compiler-Bug-found-with-abbreviated-temp/10980244

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>