Testing added for all of the Invalid concept stuff and user defined type support.
It really is just adding a bunch of tests! These are all STATIC_REQUIRE tests so they are basically guaranteed passes, but they do need to be here to make sure the invariants I am expecting are being enforced. Tests some concepts, traits objects, and the standard specializations of the getWhyInvalid and getInvalidInfo CPO objects.

The Invalid object needs testing, and I am deeply concerned with the TaggedInvalid type. I may need to approach it differently or maybe treat it differently than I was expecting to.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>