I've needed to commit for a while, but things got away from me.
So I was supposed to just add testing. Then I found the need to implement an entirely new class called PackedTuple and everything involved with that, because of an online post. I've also been working on getting my library's build system in order, and started segmenting out this library's planned functionality. I've also started the notion of adding the Task system into my main libKirHut library, so that I can use multithreading anywhere and everywhere, regardless of the build environment or existence of Qt. This means I need to have a separate thread pool implementation for when Qt isn't available, and right now I'm thinking it will be BS::thread_pool, but we'll see how things shake out for Android and iPhone builds.

I've removed the ctchecks.cpp file completely, as well as considerably streamlined how the build system works. Now tests are all in one single test executable instead of a bunch of separate ones, the libKirHutExternal library is now an "object" library so no separate .a or .lib file will be generated for it, and a lot of options have been added to include/exclude specific parts of the library based on user preference. The default, though, is to build everything. There are, of course, also documentation improvements to document the new options available in CMake for building the library differently.

A bunch of test files have been added, as well as a private header making a comeback. Honestly, again just too many changes to really name them all. I have also bumped the version to 0.1.2.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>