Mostly added more Ranges tests and updated existing tests, but some code fixes as well.
The CLI::Parser object was broken since it declared the move constructor and assignment operator as default in the header. This caused build errors where it tried to call delete on an incomplete pointer type, so the implementations were moved to the source file to fix this.

Removed unnecessary static_asserts in fromEndian and toEndian implementations, as they had been updated to use ExactUInt and ExactUIntOf previously.

Added preliminary stack trace functionality for Error type as well, which currently only works when built in C++23 mode. The documentation of the Error class continues to be massaged as I find more blatantly wrong info, but this isn't the documentation cleanup phase so that isn't what I'm concentrating on.

Renamed BullshitEnvironment to EnvironmentProblem. Will avoid names like that in the future.

Multiple fixes for the R::WordView object, along with the start of a complete test bed. I still need to test more of this implementation.

Started adding the ability to completely remove exception support and simply resort to terminating the application on exceptional conditions. In most cases, the preferred result is to simply never throw exceptions and return appropriate error codes if something goes wrong, but sometimes there is little option other than to throw or simply crash. The library should support building with both options.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>