So WordView was basically broken, as proven by the new test bed.
This is a patch that includes the new tests in TestRanges for the WordView object, and also has a fix in TestBase.cpp that fixes a build issue on GCC/Clang. The 128 bit integer tests would fail to compile on GCC/Clang because Catch2's random() generator does not support 128 bit values. The fix just skips performing this set of tests, but it would be better if a more general function could be used to generate the values, like use some kind of range adapter that the Catch2 GENERATE() macro can use as a source rather than the built-in random() generator.

More importantly, the WordView object did not properly deduct the type of string_view constructor to use with its more simplified template argument deduction guides, so a more complete solution had to be provided. This created a new StringLike concept that needs to be either pushed into a Detail namespace soon, or moved to kh/base.hpp, documented, and tested. It seems like a reasonably useful concept (basically anything that can be converted into a std::string_view is accepted), but it depends a lot on how reusable this concept is.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>