Added PackedTuple object testing, fixed the PackedTuple class (again), cleaned up the Md5Hash object, and a bit more.
Added the CharType and StringLike concepts to Base, because they were very useful for creating template argument deduction guides. There is also a StringLikeTraits object that extracts the character type of a StringLike object. Appropriate tests have been added to TestBase.cpp for these new concepts and objects.

The MaybeInv class is fundamentally broken because of all the major changes I made to the Invalid class. I will need to modify the MaybeInv appropriately for its new purpose.

There is a frexp() function I have created in the math.hpp header, but it is completely broken. I want to get the tests made first then I will fix it once they are done.

On the Md5Hash object, I corrected documentation, cleaned up some names, inlined a lot of the methods (my precious ABI!), and added a distinction between "getMd5()" and "fitMd5()". The get version will throw an exception when it cannot completely fill a buffer, whereas the fit version will just silently fill in as many bytes as fit in the given buffer. I'll need to flesh this out more later on in this class but I'm trying not to concentrate on this stuff since it doesn't really matter for this branch.

The StringLike and StringLikeType types were used in the ranges.hpp header update.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>