Mostly added the TestPrint functionality, along with some documentation updates, preparation of args.hpp for testing, and PackedTuple updates.
First, all of the testing was completed for the kh/print.hpp header file. Well, almost... there's the question of how to test the UTF-8 output on the console. It works how you would expect it to, but making sure this is the case with a unit test remains elusive to me since I cannot simply replace std::cout's rdbuf(). On Windows, the output does not go through std::cout, because that isn't wide character capable, so instead it uses the direct OS system calls after a conversion to wchar_t. If it isn't possible to unit test this given the current testing framework, I may need to come up with another way to do so, like creating a simple executable that uses the library and then testing the output of that executable.

New functionality was added to the BitField class to allow users to get reflection information about a given BitField type while also putting that information in its most natural place. The BasicPackedTuple has been updated to use this rather than the fieldDigits() function, that didn't really fit into a public interface and was only there for implementation detail anyway. Having a min() and max() function is also helpful for making it easy to know if a value can fit in a given BitField or not. Insufficient testing has been added, but I'll improve it.

A fundamental update to the API has occurred when this library is built with Qt support. The i8 to i64 and u8 to u64 types have been updated to be the equivalent to the QtTypes library qint8 to qint64 and quint8 to quint64 types respectively. This should have no real impact on any of the algorithms, since the types are all equivalent on basically every type of hardware this library has ever been used on, but I also updated the version number to 0.2.0 for this alpha version.

Some preparation has been made for the compile-time Command string checking functionality. The next major push on this testing branch is to fully test and implement the argument parsing functions in this library so that it all works as expected and correctly.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>