More features added, another version bump, and a new style guideline that I will be applying very soon.
Build options were added to make it easier for packagers to package the library, and a more accurate way to handle 128-bit support on 64 bit targets for Clang and GCC compilers. FetchContent is now optionally added if the user allows it, but it is required to stay on on Windows. It should also now be possible to require the use of the locally installed fmtlib if it can be found with find_package and is at least version 10. There was a warning that needed to be suppressed with the older version of {fmt} on Windows, but otherwise it works just as well.

Command object in argument parser is one step closer to being compile-time checked.

New Integral concept added that is effectively the same as std::integral except excludes bool. The abs() and uabs() functions have been updated to be branch free on basically all hardware, a shiftMask() function has been added that may or may not stay, the maskAt() function was renamed to shl(), and no longer accepts signed integer shifts or reverses it with negative values. Instead, the shr() function is also added for right shifts, as these follow the pattern under the "Better Shifting" proposal here: https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2025/p3793r0.html. These seem overall much better than maskAt() from the name and function, and they are also branch free on nearly all hardware that matters. The byteSwap functions have been updated to use the shift functions rather than the actual operators.

Small other fixes applied to error.hpp and global.hpp for Linux and MinGW builds.

The math.hpp header has been completely removed, it was misguided to add since this is a lot of work I do not need to worry about. If I need a math header for constexpr functions in C++20, I can add it at that time.

PackedTuple has been radically modified for the new Style Guideline for this project. Going forward, all objects will have its methods defined outside of the body of the class, including template classes and methods that are constexpr. This is because it allows for every method to have equal access to all definitions in the class object, and guarantees that the class itself is fully defined when a method is being defined. There are several situations that I have run into that have necessitated the method body being implemented outside the class, and applying this universally is both non-detrimental and actually contributes to the organization of the project. The way the Detail::FieldTraits template specialization is defined directly below the first one, and each template specialization method is defined right next to its generic template definition, is impossible with inline definition, and I like how it looks.

PackedTuple was the first class I really tried this with because it is one of the most template-heavy and hard to do, so if it can be done with this than it can be done with any of my other objects.

The print methods now attempt to detect if the passed in ostream or FILE * goes to a console, and will transcode UTF-8 to the appropriate output encoding in that case. Well, more specifically, on Windows it will use boost::nowide when it detects the user passes in std::cout/std::cerr or stdout/stderr, respectively.

The getIters() method was updated to return an explicitly named class called IterPair, which has a .begin and .end member. This may be further revised in the future, but for now it works rather well.

Appropriate testing has been added for the new shl() and shr() functions, as well as the tests for maskAt() being removed. PackedTuple's new braced initialization and std::tuple initialization now have basic tests as well, could be better though.

All tests are passing, on Windows using MinGW 13.1.0 and MSVC 19.44, and on Linux using GCC 13.3.0.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>