Small Cleanup in base.hpp and fixed tests for TestPrint.cpp and TestBase.cpp.
The TestPrint.cpp file is the most interesting, I needed to create a new EphemeralFile class that properly creates temporary files on Windows, and doing that is less-than-trivial. I may use this implementation somewhere else later on, but for now it is just part of the test infrastructure so it doesn't need docs and everything.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>