Completely removed the target_compile_definitions from CMakeLists.txt and all of the complication of the KH_EXPLICIT_TEMPLATE_EXPORT, and moved it to config.hpp.in.
This has little to do with the testing system, but the way that configuration options were passed to the library was clunky and had already grown to be larger than should be in a CMake file, but it was what worked so it stayed. I've removed all of that and now put everything into a proper config.hpp.in file, which CMake will then call configure_file() on and produce a correct config.hpp for the options provided to CMake by the user. This required unconditionally including kh/global.hpp in most headers, though that isn't really a problem.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>