Radical simplification of dependencies and CMake files, full integration of boost::nowide into libKirHutExternal, and KirHut::IO revision.
The bulk of this commit is the addition of all of the source code in boost::nowide to libKirHutExternal, which is done to eliminate the visible dependency on boost::nowide. This way a user is not required to download or worry about ensuring the nowide dependency is satisfied, as it always comes with the source code and is built-in. This makes the entire build dependency tree simpler, now libKirHut only visibly requires {fmt}, which is, itself, optional, and is part of the library's public API so that dependency should be visible.

Adding nowide to libKirHutExternal is all part of a greater effort to radically simplify the CMakeLists.txt of this project, which I am lucky to say is still not nearly as bad as a lot of library projects out there. I will continue to make concerted efforts to keep the build system as simple as possible for all of the required functionality in the build, and work to ensure that my CMake file does not step on the toes of any projects that depend on this library. This means that my library now avoids defining CMAKE_POSITION_INDEPENDENT_CODE, and instead sets the property for each target individually based on the BUILD_SHARED_LIBS variable.

How everything in KirHut::IO has been rewritten to flesh-out the Qt QIODevice output as well as using {fmt}'s text_style for output, and to accommodate {fmt} version 9, which does not include fmt::println() and friends. As such, everything had to be implemented using fmt::print() instead, to avoid the println() dependency and build with {fmt} version 9. This is a requirement for building using the Debian packaged version on LTS Ubuntu and Debian.

Multiple other smaller cleanups have been performed in this commit as well. The documentation in README.md and in other locations has been updated to reflect more current reality in the library. This is an ongoing effort, and the bulk will be done when a branch dedicated to the documentation is started.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>