Another minor version update commit, adding license compliance, v1 namespace changes, a redesign of the Argument parser is started, and a full reimplementation of the Md5Hash -> Hash::Md5 object.
Here's yet another really huge commit because for some reason I just can't commit regularly, and am doing a lot of edits all the time to this codebase. I desperately need to complete the Argument parser and build the testing around that, but I needed to also flesh out the testing on the Md5Hash object as well as rewrite it to comply with my new coding standards. While I was doing this, I ended up radically revising how the implementation works, preparing it to be completely constexpr in the future. The reimplementation is aimed at becoming a model implementation of MD5, as there isn't much other value in it, and it will eventually be moved to EigenCrypt to be the default MD5 implementation when there isn't a processor optimized version available.

PackedTuple has been given some updates, with several unnecessary functions removed and the classes involved tightened up to work in most every case. The biggest improvement I can think of is having named fields, but doing that is non-trivial to say the least. It may be something I need to wait for C++26 reflection to do properly, as well.

I added license files for all of the code I am using that isn't mine. I haven't used BS_thread_pool, mio, or toml++ yet, but they are in the library so the license should be here, and I have been making use of Boost::nowide but now I am directly using it in my source code repo so here's the license. There's also another object I am using as a reference MD5 implementation called md5_t, and unfortunately after investigating where this code even came from, it was a C++ port of some C code that is GPL code from the GNU project. This didn't matter when this project was itself licensed under a GPL, but now I need to remove this from the source tree. I will do this as soon as I have found a suitable replacement reference implementation to use.

The command line argument parser design I started with sucks, and I have been trying to come up with a better one. After some wrestling with it, I've come up with a more declarative interface that will allow easily expanding this functionality to have translated help text generation using the Options and Commands the users has defined, or simply omitting this if it is unnecessary.

In order to facilitate the heavy changes to the Hash::Md5 object's interface and source code, the endian conversion functions had to be modified to be more generic than to use pointers, and now work on std::contiguous_iterators instead.

This library will be using an alternative method of function versioning than is conventional (where every major version of the library is under a new umbrella "vX" namespace). Instead, all functions will remain under the v1 namespace until they are actually updated to a new interface, when they will then be converted to v2. This means that it is possible for even major version updates of the library to possibly be binary backward compatible with software built with the older library versions. The price of this is that there are now two effective "versions" of everything in the library: the official library release version, and the version number of the function or object itself (since each object will be updated to different version namespaces separately). Regardless, in anticipation of putting everything under a v1 inline namespace, I have provided KH_INLINE_NAMESPACE_V1 and KH_END_INLINE_NAMESPACE macros like nearly all projects do, to wrap the namespace version name. This is actually for no real reason other than Doxygen sucks: https://github.com/doxygen/doxygen/issues/5914.

I changed the interface of the Ranges namespace, and more interface changes may be coming in the future. The short namespaces are a misguided goal, given that namespace aliases exist so you never really need to do this, and it substantially increases the likelihood of a name conflict with basically anything. Longer names prevent this, and are more self-documenting, so this should be the way going forward.

Also, I renamed the linuxos identifier to gnulinux, mostly to troll Linus Torvalds.

Signed-off-by: Ezekiel Oruven <zeke@kirhut.com>